python/requirements.txt
ai21
fpdf
pymongo
python-dotenv
openpyxl
watchdog
smtplib